+++ apps/admin-panel/src/components/index.ts
@@ -2,3 +2,4 @@ export * from "./FileUpload";
 export * from "./MultiFileUpload";
 export * from "./Layout";
 export * from "./Icons";
+export * from "./Header";
